services/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
cryptography>=41.0.0
matplotlib==3.8.2
pandas==2.1.4
numpy==1.26.4
confluent_kafka==2.3.0
redis==5.0.1